+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * Sign OSX binary
+* Add support for SSE-C server-side encryption mode
 
 ### Fixed
 * Exclude packages inside the test package when installing